Refactor navigation logic and enhance user management components
Refactor navigation logic in `App.tsx` to reduce redundancy by using a dynamic array for role-specific navigation items. Create separate components for mobile and desktop user tables, enhancing maintainability and responsiveness (`MobileUserTable.tsx`, `DesktopUserTable.tsx`). Update `NewUserModal.tsx` to improve email validation and error handling, using a new utility function `validateEmail.ts`.

- Split `UserTable.tsx` into `MobileUserTable.tsx` and `DesktopUserTable.tsx`.
- Add `validateEmail.ts` for email validation across components.
- Update user creation logic to handle asynchronous errors more effectively and add a password reset feature post-user creation.

Release-Note: Improve navigation and user management, adding responsive tables and email validation.
