+++ src/pages/clients.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { Typography, Button, Box, CircularProgress } from '@mui/material';
 import User from '../models/User';
 import { PageContainer } from '@toolpad/core';
-import { getUsersByRole, deleteUser } from '../utils/firestore';
+import { getUsersByRole, deleteUser, createUser } from '../utils/firestore';
 import UserTable from '../components/UserTable';
 import NewUserModal from '../components/NewUserModal';
 import { Add } from '@mui/icons-material';
@@ -74,6 +74,7 @@ export default function ClientsPage() {
                 onClose={handleCloseModal}
                 role="client"
                 onUserCreated={handleUserCreated} // Pass the callback
+                createUser={createUser}
             />
         </PageContainer>
     );
