+++ src/utils/firestore.ts
@@ -8,6 +8,7 @@ import {
 } from 'firebase/firestore';
 import { db } from './firebaseConfig';
 import User from '../models/User';
+import { sendPasswordResetEmail, getAuth } from 'firebase/auth';
 
 export async function getUserByEmail(email: string): Promise<User> {
     try {
@@ -50,7 +51,7 @@ export async function getUsersByRole(role: string): Promise<User[]> {
     }
 }
 
-export async function createUser(user: User) {
+export async function createUser(user: User): Promise<User> {
     const response = await fetch(
         `https://us-central1-the-spa-84a52.cloudfunctions.net/createUser`,
         {
@@ -65,6 +66,10 @@ export async function createUser(user: User) {
     if (!response.ok) {
         throw new Error('Network response was not ok');
     }
+
+    // Send password reset email after user creation
+    const auth = getAuth();
+    await sendPasswordResetEmail(auth, user.email);
 
     return response; // Return response for further processing
 }
