+++ src/utils/validateEmail.ts
@@ -1,0 +1,6 @@
+function validateEmail(email: string): boolean {
+	const re = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+	return re.test(email);
+}
+
+export default validateEmail;
